refactor(PostEditor, UserInfo, uploadedImage)
- uploadImage 유틸함수로 리팩토링

#203
